--- feature/src/main/res/drawable/ic_down_arrow.xml
@@ -1,13 +1,0 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="14dp"
-    android:height="8dp"
-    android:viewportWidth="14"
-    android:viewportHeight="8">
-  <path
-      android:pathData="M2,1.966L7,6.966L12,1.966"
-      android:strokeLineJoin="round"
-      android:strokeWidth="1.5"
-      android:fillColor="#00000000"
-      android:strokeColor="#344054"
-      android:strokeLineCap="square"/>
-</vector>
