--- feature/src/main/res/drawable/ic_level_newcomer.xml
@@ -1,125 +1,0 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:aapt="http://schemas.android.com/aapt"
-    android:width="52dp"
-    android:height="80dp"
-    android:viewportWidth="52"
-    android:viewportHeight="80">
-  <path
-      android:pathData="M2.13,75.267C1.135,67.142 7.475,59.978 15.662,59.978H36.34C44.526,59.978 50.867,67.142 49.872,75.267L49.292,80H2.71L2.13,75.267Z">
-    <aapt:attr name="android:fillColor">
-      <gradient 
-          android:startX="26.001"
-          android:startY="68.222"
-          android:endX="26.001"
-          android:endY="80"
-          android:type="linear">
-        <item android:offset="0" android:color="#FFA4C5FF"/>
-        <item android:offset="1" android:color="#FF79A1E8"/>
-      </gradient>
-    </aapt:attr>
-  </path>
-  <path
-      android:pathData="M14.893,62.015a11.026,9.189 0,1 0,22.052 0a11.026,9.189 0,1 0,-22.052 0z">
-    <aapt:attr name="android:fillColor">
-      <gradient 
-          android:startX="25.919"
-          android:startY="62.015"
-          android:endX="25.919"
-          android:endY="71.204"
-          android:type="linear">
-        <item android:offset="0" android:color="#FF999999"/>
-        <item android:offset="1" android:color="#FFFFFFFF"/>
-      </gradient>
-    </aapt:attr>
-  </path>
-  <path
-      android:pathData="M23.211,66.015C23.226,65.835 23.378,65.699 23.559,65.699H28.443C28.624,65.699 28.775,65.835 28.791,66.015C28.837,66.535 28.909,67.579 28.817,68.213C28.684,69.129 28.146,70.466 28.146,70.466H23.856C23.856,70.466 23.318,69.129 23.185,68.213C23.093,67.579 23.165,66.535 23.211,66.015Z">
-    <aapt:attr name="android:fillColor">
-      <gradient 
-          android:startX="26"
-          android:startY="65.699"
-          android:endX="26"
-          android:endY="70.466"
-          android:type="linear">
-        <item android:offset="0" android:color="#FF4775C3"/>
-        <item android:offset="1" android:color="#FF4468A6"/>
-      </gradient>
-    </aapt:attr>
-  </path>
-  <path
-      android:pathData="M29.73,76.171C29.789,76.686 29.546,77.184 29.114,77.469L26.391,79.265C26.154,79.422 25.847,79.422 25.61,79.265L22.888,77.469C22.456,77.184 22.213,76.686 22.271,76.171C22.38,75.207 22.599,73.608 22.95,72.478C23.198,71.678 23.855,70.465 23.855,70.465L28.146,70.465C28.146,70.465 28.803,71.678 29.052,72.478C29.402,73.608 29.621,75.207 29.73,76.171Z">
-    <aapt:attr name="android:fillColor">
-      <gradient 
-          android:startX="26.001"
-          android:startY="71.459"
-          android:endX="26.001"
-          android:endY="76.901"
-          android:type="linear">
-        <item android:offset="0" android:color="#FF4775C3"/>
-        <item android:offset="1" android:color="#FF345690"/>
-      </gradient>
-    </aapt:attr>
-  </path>
-  <path
-      android:pathData="M17.291,70.51L11.221,62.838L18.572,63.45L25.922,65.901L18.952,70.755C18.42,71.125 17.693,71.018 17.291,70.51Z"
-      android:fillColor="#ffffff"/>
-  <path
-      android:pathData="M34.553,70.51L40.623,62.838L33.272,63.45L25.921,65.901L32.892,70.755C33.423,71.125 34.151,71.018 34.553,70.51Z"
-      android:fillColor="#ffffff"/>
-  <path
-      android:pathData="M48.972,43.757C48.987,43.416 48.994,43.072 48.994,42.726C48.994,29.833 38.697,19.381 25.996,19.381C13.295,19.381 2.998,29.833 2.998,42.726C2.998,43.082 3.006,43.436 3.021,43.787C-0.078,50.81 0.439,64.636 24.952,66.07C25.327,66.092 25.677,66.102 26.003,66.099C26.329,66.102 26.68,66.092 27.055,66.07C51.602,64.634 52.086,50.771 48.972,43.757Z"
-      android:fillColor="#FFD01A"
-      android:fillType="evenOdd"/>
-  <path
-      android:pathData="M18.847,40.711m-3.337,0a3.337,3.337 0,1 1,6.674 0a3.337,3.337 0,1 1,-6.674 0"
-      android:fillColor="#2E2216"/>
-  <path
-      android:pathData="M33.153,40.542m3.337,0a3.337,3.337 0,1 0,-6.674 0a3.337,3.337 0,1 0,6.674 0"
-      android:fillColor="#2E2216"/>
-  <path
-      android:pathData="M24.123,45.675L18.792,48.683C17.981,49.141 17.985,50.31 18.799,50.763L24.143,53.735C25.296,54.376 26.698,54.376 27.851,53.735L33.196,50.763C34.009,50.31 34.013,49.141 33.202,48.683L27.871,45.675C26.708,45.019 25.286,45.019 24.123,45.675Z"
-      android:fillColor="#EB6A2B"/>
-  <path
-      android:pathData="M25.231,48.952L19.326,49.669L25.192,50.423C25.67,50.484 26.153,50.485 26.631,50.425L32.674,49.669L26.592,48.95C26.14,48.896 25.683,48.897 25.231,48.952Z"
-      android:strokeLineJoin="round"
-      android:strokeWidth="0.953446">
-    <aapt:attr name="android:fillColor">
-      <gradient 
-          android:startX="32.674"
-          android:startY="49.49"
-          android:endX="19.326"
-          android:endY="49.49"
-          android:type="linear">
-        <item android:offset="0" android:color="#FFE36527"/>
-        <item android:offset="0.51" android:color="#FFB03E06"/>
-        <item android:offset="1" android:color="#FFE46527"/>
-      </gradient>
-    </aapt:attr>
-    <aapt:attr name="android:strokeColor">
-      <gradient 
-          android:startX="32.674"
-          android:startY="49.49"
-          android:endX="19.326"
-          android:endY="49.49"
-          android:type="linear">
-        <item android:offset="0" android:color="#FFE9682A"/>
-        <item android:offset="0.51" android:color="#FFB03E06"/>
-        <item android:offset="1" android:color="#FFE56627"/>
-      </gradient>
-    </aapt:attr>
-  </path>
-  <path
-      android:pathData="M22.846,23.417C20.431,23.477 17.533,23.224 15.739,21.945C13.246,20.168 12.172,15.902 11.769,13.678C11.623,12.87 12.136,12.123 12.941,11.964C15.351,11.488 20.321,10.808 23.013,12.592C24.669,13.689 25.651,15.845 26.232,17.911C26.813,15.844 27.795,13.688 29.452,12.591C32.146,10.807 37.12,11.488 39.532,11.964C40.337,12.123 40.85,12.87 40.704,13.678C40.3,15.902 39.227,20.167 36.732,21.945C34.822,23.306 31.66,23.504 29.162,23.401C28.656,24.514 27.534,25.287 26.232,25.287H25.768C24.472,25.287 23.356,24.521 22.846,23.417Z"
-      android:fillColor="#13B149"
-      android:fillType="evenOdd"/>
-  <path
-      android:pathData="M9.793,45.162L11.7,45.162A4.767,4.767 0,0 1,16.467 49.929L16.467,49.929A4.767,4.767 0,0 1,11.7 54.696L9.793,54.696A4.767,4.767 0,0 1,5.025 49.929L5.025,49.929A4.767,4.767 0,0 1,9.793 45.162z"
-      android:strokeAlpha="0.25"
-      android:fillColor="#FF8383"
-      android:fillAlpha="0.25"/>
-  <path
-      android:pathData="M40.3,45.162L42.207,45.162A4.767,4.767 0,0 1,46.975 49.929L46.975,49.929A4.767,4.767 0,0 1,42.207 54.697L40.3,54.697A4.767,4.767 0,0 1,35.533 49.929L35.533,49.929A4.767,4.767 0,0 1,40.3 45.162z"
-      android:strokeAlpha="0.25"
-      android:fillColor="#FF8383"
-      android:fillAlpha="0.25"/>
-</vector>
