+++ feature/vote/src/main/AndroidManifest.xml
