+++ feature/vote/src/main/java/com/example/sachosaeng/feature/vote/VoteDetailScreenUiState.kt
