+++ settings.gradle.kts
@@ -19,7 +19,6 @@ include(":app")
 include(":data")
 include(":core")
 include(":core:domain")
-include(":feature")
 include(":core:ui")
 include(":core:util")
 include(":feature:auth")
@@ -28,3 +27,4 @@ include(":feature:mypage")
 include(":feature:webview")
 include(":feature:signup")
 include(":feature:splash")
+include(":feature:vote")
